+++ backend/exif/entities/src/main/scala/backend/exif/entities/ImageExif.scala
@@ -5,7 +5,7 @@ import java.time.Instant
 case class ImageExif(cameraMake: Option[String],
                      cameraModel: Option[String],
                      lens: Option[String],
-                     focalLength: Option[Int],
+                     focalLength: Option[Float],
                      fNumber: Option[Float],
                      iso: Option[Int],
                      exposureTime: Option[String],
